+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-* no changes yet
+* Fixed: deflayer formatter no longer adds extra whitespace at end of line ([#49](https://github.com/rszyma/vscode-kanata/issues/49))
 
 ### 0.14.1